Move public types to top level, hide the shared module
