+++ crates/junction-api/src/xds/http.rs
@@ -4,9 +4,10 @@ use crate::{
     error::{Error, ErrorContext},
     http::{
         HeaderMatch, Method, PathMatch, QueryParamMatch, Route, RouteMatch, RouteRetry, RouteRule,
-        RouteTimeouts,
+        RouteTimeouts, WeightedTarget,
     },
-    shared::{Duration, Regex, Target, WeightedTarget},
+    shared::{Duration, Regex},
+    Target,
 };
 use xds_api::pb::{
     envoy::{
@@ -549,9 +550,65 @@ where
     }
 }
 
+impl WeightedTarget {
+    pub(crate) fn to_xds(targets: &[Self]) -> Option<xds_route::route_action::ClusterSpecifier> {
+        match targets {
+            [] => None,
+            [target] => Some(xds_route::route_action::ClusterSpecifier::Cluster(
+                target.target.xds_cluster_name(),
+            )),
+            targets => {
+                let clusters = targets
+                    .iter()
+                    .map(|wt| xds_route::weighted_cluster::ClusterWeight {
+                        name: wt.target.xds_cluster_name(),
+                        weight: Some(wt.weight.into()),
+                        ..Default::default()
+                    })
+                    .collect();
+
+                Some(xds_route::route_action::ClusterSpecifier::WeightedClusters(
+                    xds_route::WeightedCluster {
+                        clusters,
+                        ..Default::default()
+                    },
+                ))
+            }
+        }
+    }
+
+    pub(crate) fn from_xds(
+        xds: Option<&xds_route::route_action::ClusterSpecifier>,
+    ) -> Result<Vec<Self>, Error> {
+        match xds {
+            Some(xds_route::route_action::ClusterSpecifier::Cluster(name)) => Ok(vec![Self {
+                target: Target::from_cluster_xds_name(name).with_field("cluster")?,
+                weight: 1,
+            }]),
+            Some(xds_route::route_action::ClusterSpecifier::WeightedClusters(
+                weighted_clusters,
+            )) => {
+                let clusters = weighted_clusters.clusters.iter().enumerate().map(|(i, w)| {
+                    let target =
+                        Target::from_cluster_xds_name(&w.name).with_field_index("name", i)?;
+                    let weight = crate::value_or_default!(w.weight, 1);
+
+                    Ok(Self { target, weight })
+                });
+
+                clusters
+                    .collect::<Result<Vec<_>, _>>()
+                    .with_fields("weighted_clusters", "clusters")
+            }
+            Some(_) => Err(Error::new_static("unsupporetd cluster specifier")),
+            None => Err(Error::new_static("missing cluster specifier")),
+        }
+    }
+}
+
 #[cfg(test)]
 mod test {
-    use crate::shared::ServiceTarget;
+    use crate::ServiceTarget;
 
     use super::*;
