+++ crates/junction-core/src/error.rs
@@ -1,6 +1,6 @@
 use std::borrow::Cow;
 
-use junction_api::shared::Target;
+use junction_api::Target;
 
 /// A `Result` alias where the `Err` case is `junction_core::Error`.
 pub type Result<T> = std::result::Result<T, Error>;
